feat(wezterm): add WebGpu as front end and ensure Wayland is enabled for better graphics support fix(zprofile): improve PATH variable management by checking for directory existence before adding chore(zprofile): comment out problematic DBUS environment variable for vim zathura integration feat(configuration.nix): add initial NixOS configuration with essential packages and services feat(hardware-configuration.nix): include generated hardware configuration for system setup

+++ .zprofile
@@ -5,7 +5,6 @@ if [ -e $HOME/.nix-profile/etc/profile.d/nix.sh ]; then . $HOME/.nix-profile/etc
 
 # Check if the OS is Linux or macOS
 if [[ "$OS_TYPE" == "Linux" ]]; then
-    test -e "/home/linuxbrew/.linuxbrew/bin/brew" && eval "$(/home/linuxbrew/.linuxbrew/bin/brew shellenv)"
     export PINENTRY_PROGRAM="$(which pinentry)"
 elif [[ "$OS_TYPE" == "Darwin" ]]; then
     eval "$(/opt/homebrew/bin/brew shellenv)"
@@ -19,17 +18,52 @@ export GPG_TTY=$(tty)
 if [ -z "$TMUX" ]; then
 
   # MacPorts Installer addition on 2023-05-26_at_19:45:00: adding an appropriate PATH variable for use with MacPorts.
-  export PATH="/opt/local/bin:/opt/local/sbin:$PATH"
+  if [ -d /opt/local/bin ]; then
+    export PATH="$PATH:/opt/local/bin"
+  fi
 
-  export PATH="$HOME/.cargo/bin:$PATH"
+  if [ -d /opt/local/sbin ]; then
+    export PATH="$PATH:/opt/local/sbin"
+  fi
 
-  export PATH="$HOME/.local/bin:$PATH"
+  if [ -d /$HOME/.cargo/bin ]; then
+    export PATH="$PATH:$HOME/.cargo/bin"
+  fi
 
-  export PATH="$HOME/.npm-global/bin:$PATH"
+  if [ -d $HOME/.npm-global/bin ]; then
+    export PATH="$PATH:$HOME/.npm-global/bin"
+  fi
 
-  export PATH="/opt/homebrew/bin:/opt/homebrew/sbin:/Users/$USER/.nix-profile/bin:/etc/profiles/per-user/$USER/bin:/run/current-system/sw/bin:/nix/var/nix/profiles/default/bin:$PATH"
+  if [ -d /opt/homebrew/bin ]; then
+    export PATH="$PATH:/opt/homebrew/bin"
+  fi
+  if [ -d /opt/homebrew/sbin ]; then
+    export PATH="$PATH:/opt/homebrew/sbin"
+  fi
 
-  export PATH="$PATH:/Users/$USER/.local/bin"
+  if [ -d $HOME/.nix-profile/bin ]; then
+    export PATH="$PATH:$HOME/.nix-profile/bin"
+  fi
+
+  if [ -d /etc/profiles/per-user/$USER/bin ]; then
+    export PATH="$PATH:/etc/profiles/per-user/$USER/bin"
+  fi
+
+  if [ -d /run/current-system/sw/bin ]; then
+    export PATH="$PATH:/run/current-system/sw/bin"
+  fi
+
+  if [ -d /nix/var/nix/profiles/default/bin ]; then
+    export PATH="$PATH:/nix/var/nix/profiles/default/bin"
+  fi
+
+  if [ -d $HOME/.local/bin ]; then
+    export PATH="$PATH:$HOME/.local/bin"
+  fi
+
+  if [ -d $HOME/.config/tmux/scripts ]; then
+    export PATH="$PATH:$HOME/.config/tmux/scripts"
+  fi
 
   if [ -d /Applications/Docker.app/Contents/Resources/bin ]; then
     export PATH="$PATH:/Applications/Docker.app/Contents/Resources/bin"
@@ -37,13 +71,19 @@ if [ -z "$TMUX" ]; then
 
   # Adding latexmk
   # TODO install latexmk after installing mactex using 'tlmgr install latexmk'
-  export PATH="/Library/TeX/texbin:$PATH"
+  if [ -d /Library/TeX/texbin ]; then
+    export PATH="$PATH:/Library/TeX/texbin"
+  fi
 
   # Added to enable lib in Cargo
-  export LIBRARY_PATH="$HOME/.nix-profile/lib:$LIBRARY_PATH"
+  if [ -d $HOME/.nix-profile/lib ]; then
+    export LIBRARY_PATH="$LIBRARY_PATH:$HOME/.nix-profile/lib"
+  fi
 
   # MacPorts Installer addition on 2023-05-26_at_19:45:00: adding an appropriate MANPATH variable for use with MacPorts.
-  export MANPATH="/opt/local/share/man:$MANPATH"
+  if [ -d /opt/local/share/man ]; then
+    export MANPATH="$MANPATH:/opt/local/share/man"
+  fi
   # Finished adapting your MANPATH environment variable for use with MacPorts.
 fi
 
@@ -58,8 +98,9 @@ alias vimv='vimv -e vim'
 alias vi="nvim -c 'if filereadable(\"Session.vim\") | source Session.vim | endif'"
 alias vim="nvim -c 'if filereadable(\"Session.vim\") | source Session.vim | endif'"
 
-# Added DBUS for vim zathura integration
-export DBUS_SESSION_BUS_ADDRESS="unix:path=$DBUS_LAUNCHD_SESSION_BUS_SOCKET"
+# # Added DBUS for vim zathura integration
+# # Creates problems in the nixos login
+# export DBUS_SESSION_BUS_ADDRESS="unix:path=$DBUS_LAUNCHD_SESSION_BUS_SOCKET"
 
 alias h='cd ~'
 alias o='fzf -m | xargs -I % open %'
@@ -71,11 +112,15 @@ alias ta='tmux attach'
 alias c='dir=$(fzf | xargs -I {} dirname "{}") && cd "$dir"'
 alias t="tree -d -L 7| grep --color="never" -E '── \d\d-'"
 
-alias upgrade_apps='brew upgrade --cask --no-quarantine --greedy'
+if command -v brew &> /dev/null; then
+  alias upgrade_apps='brew upgrade --cask --no-quarantine --greedy'
+fi
 
-alias flushdns='sudo dscacheutil -flushcache; sudo killall -HUP mDNSResponder'
-alias finderhide='defaults write com.apple.finder CreateDesktop -bool false; killall Finder'
-alias findershow='defaults write com.apple.finder CreateDesktop -bool true; killall Finder'
+if [[ "$OS_TYPE" == "Darwin" ]]; then
+  alias flushdns='sudo dscacheutil -flushcache; sudo killall -HUP mDNSResponder'
+  alias finderhide='defaults write com.apple.finder CreateDesktop -bool false; killall Finder'
+  alias findershow='defaults write com.apple.finder CreateDesktop -bool true; killall Finder'
+fi
 
 alias passg='pass generate'
 alias passc='pass edit'
@@ -90,12 +135,14 @@ function dump_files() {
 }
 
 # check if bat is installed
-if command -v bat &> /dev/null
-then
+if command -v bat &> /dev/null; then
     alias cat='bat'
 fi
 
-wallpaper () { automator -i "${1}" ~/.Workflows/SetWallpaper.workflow }
+if command -v automator &> /dev/null; then
+  wallpaper () { automator -i "${1}" ~/.Workflows/SetWallpaper.workflow }
+fi
+
 wttr () { curl wttr.in/$1 }
 cheat() { curl cheat.sh/$1 }
 
@@ -109,10 +156,10 @@ _fzf_complete_pass() {
   )
 }
 
-# Check if the `pass` command successfully retrieves the API key
-if api_key=$(pass show api-key/anthropic 2>/dev/null); then
-    export ANTHROPIC_API_KEY="$api_key"
-fi
-if api_key=$(pass show api-key/oco 2>/dev/null); then
-    export OCO_API_KEY="$api_key"
-fi
+# # Check if the `pass` command successfully retrieves the API key
+# if api_key=$(pass show api-key/anthropic 2>/dev/null); then
+#     export ANTHROPIC_API_KEY="$api_key"
+# fi
+# if api_key=$(pass show api-key/oco 2>/dev/null); then
+#     export OCO_API_KEY="$api_key"
+# fi
